Up and Down
Now you can move blocky up and down!
(note: diagonal move is a thing. And I'm working on it)

Now the drawing is on a timer instead of only when you move (prob will change later)

Other than that there's not much else to it
( warning: code may look bad)